refactor: remove redundant comments and update `quantityProduced` type in production DTO
- Eliminate unnecessary inline comments in `ProductionService` for improved code readability.
- Change `quantityProduced` in `ProductProductionResponse` from `Long` to `BigDecimal` to enhance precision.

+++ src/main/java/com/tiagosune/inventory_api/dto/production/ProductProductionResponse.java
@@ -9,6 +9,6 @@ public class ProductProductionResponse {
 
     private Long productId;
     private String productName;
-    private Long quantityProduced;
+    private BigDecimal quantityProduced;
     private BigDecimal totalValue;
 }
